Fix the config.json file loading empty on live environments
When deployed the release ended up telling the client to serve fake
data instead of live data.  This seems to be due to the `revalidate`
being set to a non zero value.  This ultimately controls how often the
server will refresh the specific content.

The fix, for now, is to stop the server from caching the `config.json`
contents, and instead allow the client to cache it as necessary
based on the returned `Cache-Control` header.