Modify Dockerfile to remove config.json
When a `config.json` file is present in the `public` folder of the
`block-explorer` at the path
`/app/packages/block-explorer/public/config.json` it will serve that
file over the `config.json/route.ts` file that should be serving it
instead.

This modification ensures that the file is not present in the resulting
image.