Fix asset loading for storybook on github pages
With the Block Explorer now being public the github page's URL has
been modified to be public facing.  This has placed the storybook
under another path.  The images and fonts being loaded were
assumed to be able to be loaded from the root directory of the
page.  However, with the new URL being under a sub-path, this is
definitely not guaranteed to be the case.  As a result, the images
and fonts are not loading in the hosted Storybook page.

This change relocates and reorganizes the images and fonts being
utilized into relative imports that are used and referenced in the
react files and css.