Fix StoryBookPathResolver navigation on GitHub Pages
The StorybookPathResolver works well when served under a host
that is not under a sub-path.  With the change of the page host to
be a subpath in GitHub Pages the path resolver's navigation results
in errant navigation references.

This fixes the issue by making all navigation a relative modification
instead of an absolute path modification.