Fix the Storybook Stories for Staking Modal
The Storybook Stories for the Staking Modal's individual states
isn't displaying the different states like it once was.  This has to
do with a somewhat recent change made in the separation of the
modal's different rendering states.  When this was done the
specific React Contexts for the relative modals became much
more focused and targeted overwriting upstream React Contexts
that were being utilized to provide the Story states.

This fix was to separate these newer React Contexts as a separate
component wrapping the content by default, so the current
behavior is preserved, but creating a separate exported component
that can be referenced byu the story who provides its own
React Contexts already.