Add tooltips to table headings, remove rank column [Delegation UI]
Based on GTM feedback given by @clu8

Asana Task: https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212346803455540?focus=true

Add tooltips to Table Heading for Delegation UI

In order to provide better information as to the meaning / details
of the specific columns used for the validators table, the suggestion
has been put forth to add tooltip where appropriate.  These tooltips
should be able provide extra details / insights as to the intention or
meaning of each column, or how the value is derived / calculated.

Remove Rank column from validators table for Delegation UI

The Rank column essentially just enumerates the validator list
by the amount of total Stake a validator has.  This is adding a
bit too much noise as this value is already readily available within
the table, and the rank is just stating the same thing as the other
column.  As such, we don't want to include it in the table as a
directly displayed field, or as a sortable column.

That being said, the `rank` still does show up in other elements /
aspects of the UI to help denote / disambiguate validators from
each other.  It is still utilized in the mobile view, and the the select
validators modal after pressing "+ Delegate"