Remove APY Chip from Validator Information
Based on GTM feedback provided by Lucca

Asana Task: https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212346803455545?focus=true

Lucca has provided the feedback to be consistent about utilizing
`APR` versus `APY`.  Initially the intention was to replace the
`APY` chip value with an `APR` chip value for the specific
validator.  However, due to the potential complication, caveats,
and considerations needed to be conveyed for this case, the
conversation resolved to just dropping this value per the validator.