Add APR retrieval to Staking UI Service
Asana Task: https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212346803455546?focus=true

The Staking UI Service is now able to return the APR.  As a result
some of the types need to be updated in order to retrieve the
results accurately.

Additionally, we can remove the hard-coded APR value and replace
it with the value retrieved from the Service itself.