Update `RatioEncoder` to support rational
The `RatioEncoder` only encodes things as a floating point by
default which is somewhat divorced from the two different types
of representations supported / expected.  In order to make things
fall more in line with expectation, if the underlying value of the
Ratio is a `Rational` value instead of a floating point, it should
encode to that representation in order to preserve accuracy.