Fix update logic for local APR population
When applying the update for the active node set's NewEpoch
event, the updated APR is being dropped in favor of the previous
APR.  This causes us to lose track of the new APR, and makes our
UI not update accordingly.

This change fixes the underlying bug caught by @jbearer in a
code review.