Add Validator Node Metadata
The Delegation UI calls for the display of identifiable node data
within the various aspects of the design.  In order to facilitate this
the Staking UI has been modified to return the metadata when it
is available.  This data is optional or may be in various states of
retrieval making all of this data volatile and very prone to being
missing.  However, if it is available, we should utilize it in order to
assist the user in making the nodes readily identifiable.

For The validator names, we prioritize displaying the Validator
Node name wherever possible.  However it if is not possible to
display the Validator's name, we will fall back on the Address
of the Validator instead.

For the Validator Image, we prioritize utilizing the asset URL
provided to us by them.  We make a best effort to return the
specific URLs given to us falling back to any of them if any
are provided.  If no assets are provided, we instead opt to
utilize an Ethereum blockies image instead, since they are
well known and consistently verifiable at a glance.