Fix test failure due to blockies
Ethereum Blockies utilizes the `canvas` HTML element in order to
create the blockies images for the given address.  The `canvas`
element does not exist in the `vitest` environment, and as a result
a mock is needed in order to make the interactions pass without
error.

To fix this we utilize `vitest-canvas-mock` and configure it for
usage in the `vitest` environment.