Add OptionalCodec and pass for Staking UI Service
The Staking UI Service is going to be omitting keys rather than encoding
the value `null` in the future.  As a result, we need to be able to support
decoding the values when the keys are present, and `null`, or when they
are omitted entirely.

Additionally, in order to seek parity with the implementation, when we
are storing `null`, we would also like to omit them from the object
representation.

This approach calls for a `Codec` that is able to handle the encoded
values, `null`, and `undefined` in the same place.