Refactor source data contexts for Delegation UI
The Delegation UI is currently re computing its layout every time it updates.
This is somewhat expected, as every time a new L1 Block is added, the underlying
list of validators is modified with new representations.  Additionally, every time the
Espresso Block Changes the criteria for filtering considerations is also changed.
These two combined result in the data coming from the underlying table being
modified.  Add in the sort after the filter, and you have yet another case for the
list source being continually modified  with an update.

That being said, we might be able to reduce these layout needs by considering
a simpler object rather than the `NodeSetEntry` itself.  The `address` should
be more than sufficient for finding these entries when we need them.  As a result
the `React` `Context`s that drive these updates have been modified to specify
the `Address` itself instead.

By doing this the hope is that we can eliminate or reduce a lot of large tree
re-renders / re-evaluations.