Fix missed ValidatorSelectionEnum changes
When we modified the layout to utilize strings for quicker comparisons, we also
modified the `ValidatorSelectionEnum` to utilize these strings rather than the
`ArrayBuffer`s we were using.  However, in our changes we missed a few key
points where the validator address was actually being assigned.

This change rectifies these misses by addressing the missed assignments.