Improve Component Reevaluation Performance
When our sub components attempt to filter or sort the list of validator nodes, we
end up with a new `Array` that is not strictly equal to the previous array.  Since
any criteria for the source, or adjustment of these lists and criteria will trigger a
reevaluation.  This evaluation in some cases is triggered as often as every
second (for claimable rewards and exits).  Otherwise these re evaluations will
occur whenever an espresso block is updated, an l1 blocks is updated, or when
the user interacts with the UI to sort or filter things.

Due to this constant evaluation we very often end up with results that do not
change the contents of the driving array in a meaningful way, yet these will
still cause full downstream component re-evaluations.

With the change to the driving source data being addresses entirely, this
can be easily mitigated.  We simply need to compare the contents of the source
list to that of the new evaluation.  If they are different, then no updates need to
be preformed, and the source list doesn't actually need to be modified.