Replace Feather Icons with Sharp Icons
In a discussion with the Design Team, inconsistent icon size was brought up
and it was discovered that the icons being used weren't actually from the
Feather Icons set as the Design guide specified.  But rather the icons being
utilized were from the Sharp Icons set by Streamline.

This change replaces the Feather Icons with the Shap Icons, and updates the
references and styles where relevant.