Update Contract Implementations
There have been recent changes in the `RewardClaims` contract that allow for the
user to retrieve the total claimed value.  Additionally, there have been some
adjustments to the `StakeTableV2` contract that need to be accounted for.  The
`StakeTableV2` contract can also have some additional write functions added to
it for reference.

Add LightClient Contract support

While inspecting the decaf environment is was noticed that the claimRewards
gas estimate was failing.  This was due to invalid input.  After speaking with
@sveitser about the failure, it was determined that the claimRewards function
can only reference the reward claim input from the finalized state stored in the
light client contract.  This means we need to retrieve the reward claim input
utilizing the block height stored in the finalized Light Client contract.  This
additionally means that we need to interact with the Light Client contract
in order to make this happen.

We add the Light Client contract and modify the rewards claim input retrieval
to utilized the finalized block height from it to make it work.

Cleanup some redundant code patterns in Smart Contracts

The Smart contracts all start with the same function call prefix for the arguments
passed.  This ends up being a lot of wasted effort and time when it comes to
managing the implementation.  As a result, we simplify the call and create some
utility methods to make this more convenient.