Revise ESP Input behavior
The changes introduced trying to ensure only a valid
value is displayed for any ESP Input value, the UX and
editing behavior has been left wanting.  After discussions
with Mariana, we observed Polygon's Staking UI and decided
to try and aim for a similar experience.

In this way we can specify the modifications that are and
are now allowed based on the key down events, and the paste
event.  We only allow for a valid input in these cases, provided
that the parsed value makes sense.

This approach helps to eliminate a whole class of erroneous
inputs by prevents the user from inputting them in the first
place.  This leads the user to input a valid format themselves
thereby eliminating the disconnect between the interpreted
value, and the input value.

In this way what you see is what you get.  Though, we have
chosen to eliminate the grouping separator for convenience
so the user does not need to worry about inputting that value.

With this new approach we also separate the currency code
from the input entirely, so we don't need to worry about the
cursor position relative to the Currency Code itself.  This
makes the interaction area a lot more obvious, while still
respecting the currency code position for the user.

The typing input tests have also been reworked to anticipate this change, and verify the correct formatting behavior.

Additionally, we are also able to support simple number
substitution maps for convenience.