Adjust font for node search
Based on feedback from outspace concerning design
adherence.

Asana Task https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212305233315223?focus=true

The font family currently in use by the node search box is
not "IBM Plex Mono".  This is contrary to the current design.

It seems that the `input` styling doesn't automatically inherit
the font-family that is in use for the page.  As a result, the
fix for this is to just set the font-family explicitly on the
input or to inherit the page's font family.

Adjust icon position for `+ Delegate` modal

The Magnifying glass position in the search input in the select
a validator modal is currently not centered.  This is due to the
icon size changing due to the icon switch.

Change input type of Node Search to type "search"

The Search box works absolutely fine for the node search.
However, in order to provide better context clues as to its
usage and intention for both mobile, and accessibility needing
users, the type is better served if it is a "search" instead of a
normal "text".

This change replaces the type of the search box for node
from "text" to "search".