Update `+ Delegate` button start with selection
Based on feedback from Outpace

Asana Task: https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212311464235738?focus=true

Based on feedback from Outpace, we want to try and prioritize the
validator selection modal to always have a selection.  Now this isn't
always possible, especially if the node list isn't populated.  Instead
we make a best effort, we attempt to select the "first" validator in the
address list when sorted by stake.  If there is an address to be found,
we will set that as the initial selection, otherwise we will ensure that
we denote that "NoValidatorSelected" is the option, as it will reflect
reality and may stop potential errors from occuring.