Replace table column for pending exits and undelegations
Based on feedback from Marian and Outpace

Asana Task: https://app.asana.com/1/1208976916964769/project/1211758952171173/task/1212480546827260?focus=true

An initial review / conversation with Mariana resulted in some confusion
concerning the `Total Staked` column in the table for the pending exits
and pending undelegations. Mariana expected this value to reflect what
the active wallet's delegated stake to the validator is.

This prompted a question to Outpace regarding the intention of this column
and its contents.  After conferring it was indeed intended to represent
the active wallet's current stake that was available and pending withdrawal.
It was also determined that the column title should be renamed in order
to alleviate any ambiguity.

To support this we need to be able to sort by pending exits, and pending
undelegations.  Additionally, we need to modify the tables so that they
refer to different sort column criteria, creating the need for distinct
Validator Table headers between the two.  They only different by a single
column, so this can potentially be improved to reduce code duplication
in the future.