Add or improve tooltips showing calculations
The current tooltip for "Epoch ends in" is extremely lacking.  It was
intended to display the current block, but instead only displays the
epoch value again.

Additionally, the code for deriving the `blocksPerEpoch` from
`EpochAndBlock`  is not implemented correctly, and should take
implementation details from the relevant code in the LightClient
smart contract, or in the rust code base for `espresso-network`.

Add tooltip to Estimated Transaction Fee

The Estimated fee for submitting a transaction to the L1 is comprised of
a few different source numbers.  It can be useful or even beneficial to be
able to expose this raw values to the curious user to verify the state and
price that goes into determining this value.

Add Custom Currency Code Number Format

When attempting to utilize `"Gwei"` as the currency code for formatting
Gwei based currency denominations, it was noticed that the code was
not acceptable due to not being a standard ISO4217 currency code.  We
still want to be able to represent `"Gwei"` and as such we require the
need / capability to represent custom currency codes in a custom
Intl.NumberFormat intended to display currencies.