Add a Story for New Delegation Staking with value lower than minimum
It is beneficial to be able to consider all of the potential states that are reachable
for the staking modals.  As a result, since we have a no potential error due to the
user not inputting a value that exceeds the minimum, we want a Story to showcase this
case.