Fix block and epoch inconsistency in Fake Data
The Fake Data demo is failing an assertion when transitioning between
epochs.  This is due to an updated calculation for determining the epoch
given a `block height` and `blocks per epoch` value.  When this calculation
for the assertion was updated, the static helper function that is utilized to
populate the fake data wasn't.