Fix blocks per epoch information sourcing
The blocks per epoch is not guaranteed to be uniquely derivable given just
an `epoch` and a `block height`.  It seems that this may be less ambiguous
the larger the epoch becomes, but for the fake data source the block
number and epoch remain small enough to cause issues.  It has been
determined to not be worth the hassle of being incorrect about the blocks
per epoch value when we can simply retrieve it from the Light Client V2
contract.

As a result, ths blocksPerEpoch data has been provided via a Context that
is sourced from the LightClientV2 contract usage.  previous calculation
is only utilized as a fallback, and may be removed at a later point.