Remove "Claim In" Table heading
Based on feedback from GTM team

The validator table field "Claim in" is causing a bit of confusion as to
what it is for and in terms of what it pertains to.  Even with the addition
of the tooltip, the need and purpose of the column is not understood
by the larger GTM team and likely the end user in general.

After a visual demo of the purpose of the `Claim In` column, and what
it indicates, it was advised by Lucca that it would likely be better if we
just made the pending exits / claims show up in their own table even
when they ware not currently claimable.  From there we could decide
to show the count down timer or the "Withdraw" button as
appropriate.